cambios 03  de junio 10:34  am

+++ src/components/ProductoComponents/ProductosAdmin/CardProductoAdmin.jsx
@@ -2,13 +2,14 @@ import "./CardProductoAdmin.css";
 import { useProductos } from "../ProductosContext/ProductoProvider";
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
-
+import Swal from "sweetalert2";
 import { deleteImagesProducto } from "../../../firebase/productoStorage";
 import { FormUpdateProductoAdmin } from "./FormUpdateProductoAdmin";
 import { Modal } from "../../ModalComponents/Modal";
 
 export function CardProductoAdmin({ producto }) {
-  const { deleteProducto, getImgPortadaProducto, getAllImagesProduct } =  useProductos();
+  const { deleteProducto, getImgPortadaProducto, getAllImagesProduct } =
+    useProductos();
 
   const [urlImagen, setUrlImagen] = useState("");
 
@@ -26,8 +27,40 @@ export function CardProductoAdmin({ producto }) {
       const imagenesProducto = await getAllImagesProduct(producto.id);
       const urls = imagenesProducto.map((imagen) => imagen.url);
 
-      await deleteImagesProducto(urls);
-      await deleteProducto(producto.id);
+      // Mostrar SweetAlert2 de confirmación
+      const confirmResult = await Swal.fire({
+        icon: "warning",
+        title: "Eliminar producto",
+        text: `¿Estás seguro de eliminar el producto "${producto.nombre}"?`,
+        showCancelButton: true,
+        confirmButtonText: "Eliminar",
+        cancelButtonText: "Cancelar",
+      });
+
+      if (confirmResult.isConfirmed) {
+        const status = await deleteProducto(producto.id);
+
+        if (status === 200) {
+          await deleteImagesProducto(urls);
+          Swal.fire({
+            icon: "success",
+            title: "Éxito",
+            text: "El producto se eliminó correctamente",
+          });
+        } else if (status === 404) {
+          Swal.fire({
+            icon: "error",
+            title: "Error",
+            text: "El producto no se encontró o no existe",
+          });
+        } else if (status === 500) {
+          Swal.fire({
+            icon: "error",
+            title: "Error",
+            text: "Hubo un error en el servidor",
+          });
+        }
+      }
     } catch (error) {
       console.error(error);
     }
@@ -55,8 +88,14 @@ export function CardProductoAdmin({ producto }) {
       handleCloseModal(); // Cerrar la ventana modal si el formulario se ha enviado correctamente
     }
   }, [formularioEnviado]);
+  
   return (
     <div className="card-producto">
+      {producto.stock === 0 && (
+        <div className="agotado-label">
+          <h4>Agotado</h4>
+        </div>
+      )}
       <Link
         to={`/dashAdmin/productos/detalles/${producto.id}`}
         style={{ textDecoration: "none" }}
